Move es tests into their specific search apps
